git-svn-id: https://svn.code.sf.net/p/kaldi/code/sandbox/dan@1800 5e6a8d80-dfce-4ca6-a32a-6e07a63d50c8

+++ egs/wsj/s5/utils/reverse_lm.sh
@@ -1,15 +1,12 @@
 #!/bin/bash
 
 # Copyright 2012  Brno University of Technology (Author: Mirko Hannemann)
+# JHU (Author: Dan Povey)
 # Apache 2.0
 
 # configuration section
-lmdir=data/local/nist_lm  # original arpa file
-lm_suffix=tgpr_5k         # used for directory name
 tmpdir=data/local/lm_tmp  # only for OOVs and checks
-datadir=data/lang.reverse # for lexicon transducer and word lists
 lexicon=data/local/lang_tmp.reverse/lexicon.txt # only for checks
-lm=
 # end config section
 
 mkdir -p $tmpdir
@@ -19,33 +16,29 @@ echo "$0 $@"  # Print the command line for logging
 [ -f ./path.sh ] && . ./path.sh; # source the path.
 . parse_options.sh || exit 1;
 
-if [ $# != 1 ]; then
-   echo "Usage: local/reverse/wsj_reverse_lm.sh [options] <lm-suffix>"
-   echo "... where <data-dir> is copied into the new decoding data dir"
-   echo "e.g.: local/reverse/wsj_reverse_lm.sh tgpr_5k"
+if [ $# != 3 ]; then
+   echo "Usage: utils/reverse_lm.sh [options] <arpa-gz-file> <lang-dir> <out-dir>"
+   echo "e.g.: utils/reverse_lm.sh data/local/nist_lm/lm_tgpr_5k.arpa.gz data/lang.reverse data/lang_test_tgpr_5k.reverse"
+   echo "... where files from <lang-dir> are copied into <out-dir>"
    echo "options:"
-   echo " --lm <arpa-gz-file>        if non-standard lm is used"
-   echo " --datadir <data-dir>       origin of lexicon transducer (data/lang.reverse/)"
    echo " --lexicon <lexicon-file>   reversed lexicon (only for checks)"
    exit 1;
 fi
 
-lm_suffix=$1
-if [ -z "$lm" ]; then
-  lm=$lmdir/lm_${lm_suffix}.arpa.gz
-fi
-test=data/lang_test_${lm_suffix}.reverse # output directory
+lm=$1 # gzipped arpa file
+langdir=$2
+outdir=$3 # output directory
 
 # create the corresponding FST for the language model
 # and the corresponding lang_test_* directory.
 
-echo Preparing reverse language model from $lm into $test
+echo Preparing reverse language model from $lm into $outdir
 echo "Finding OOVs and strange silences"
-mkdir -p $test
+mkdir -p $outdir
 for f in phones.txt words.txt L.fst L_disambig.fst phones/; do
-  cp -r $datadir/$f $test
+  cp -r $langdir/$f $outdir
 done
-gunzip -c $lm | utils/find_arpa_oovs.pl $test/words.txt  > $tmpdir/oovs_${lm_suffix}.txt
+gunzip -c $lm | utils/find_arpa_oovs.pl $outdir/words.txt  > $tmpdir/oovs.txt
 
 # grep -v '<s> <s>' because the LM seems to have some strange and useless
 # stuff in it with multiple <s>'s in the history.  Encountered some other similar
@@ -55,23 +48,23 @@ gunzip -c $lm | utils/find_arpa_oovs.pl $test/words.txt  > $tmpdir/oovs_${lm_suf
 gunzip -c $lm | \
   grep -v '<s> <s>' | \
   grep -v '</s> <s>' | \
-  grep -v '</s> </s>' > $test/forward.arpa
+  grep -v '</s> </s>' > $outdir/forward.arpa
 echo "Mapping ARPA to reverse ARPA"
-python local/reverse_arpa.py $test/forward.arpa > $test/reverse.arpa
-arpa2fst $test/reverse.arpa | fstprint | \
+python utils/reverse_arpa.py $outdir/forward.arpa > $outdir/reverse.arpa
+arpa2fst $outdir/reverse.arpa | fstprint | \
   grep -v "230258.5" | \
-  utils/remove_oovs.pl $tmpdir/oovs_${lm_suffix}.txt | \
-  utils/eps2disambig.pl | utils/s2eps.pl | fstcompile --isymbols=$test/words.txt \
-    --osymbols=$test/words.txt  --keep_isymbols=false --keep_osymbols=false \
-    | fstrmepsilon > $test/G_org.fst
+  utils/remove_oovs.pl $tmpdir/oovs.txt | \
+  utils/eps2disambig.pl | utils/s2eps.pl | fstcompile --isymbols=$outdir/words.txt \
+    --osymbols=$outdir/words.txt  --keep_isymbols=false --keep_osymbols=false \
+    | fstrmepsilon > $outdir/G_org.fst
 #--arc_type=log
 
 echo "Push weights to make it stochastic (log semi-ring)"
 # delta must be very small otherwise weight pushing won't succeed
-#fstpush --push_weights=true --push_labels=true --delta=1E-7 $test/G_log.fst >$test/G_log_pushed.fst
-fstpushspecial --delta=1E-5 $test/G_org.fst >$test/G.fst
+#fstpush --push_weights=true --push_labels=true --delta=1E-7 $outdir/G_log.fst >$outdir/G_log_pushed.fst
+fstpushspecial --delta=1E-5 $outdir/G_org.fst >$outdir/G.fst
 
-fstisstochastic $test/G.fst
+fstisstochastic $outdir/G.fst
 # The output is like:
 # 9.14233e-05 -0.259833
 # we do expect the first of these 2 numbers to be close to zero (the second is
@@ -83,14 +76,16 @@ fstisstochastic $test/G.fst
 # Checking that G has no cycles with empty words on them (e.g. <s>, </s>);
 # this might cause determinization failure of CLG.
 # #0 is treated as an empty word.
-mkdir -p $tmpdir/g
-awk '{if(NF==1){ printf("0 0 %s %s\n", $1,$1); }} END{print "0 0 #0 #0"; print "0";}' \
-  < "$lexicon"  >$tmpdir/g/select_empty.fst.txt
-fstcompile --isymbols=$test/words.txt --osymbols=$test/words.txt $tmpdir/g/select_empty.fst.txt | \
-  fstarcsort --sort_type=olabel | fstcompose - $test/G.fst > $tmpdir/g/empty_words.fst
-fstinfo $tmpdir/g/empty_words.fst | grep cyclic | grep -w 'y' && 
-echo "Language model has cycles with empty words" && exit 1
-rm -r $tmpdir/g
 
+if [ -f $lexicon ]; then
+  mkdir -p $tmpdir/g
+  awk '{if(NF==1){ printf("0 0 %s %s\n", $1,$1); }} END{print "0 0 #0 #0"; print "0";}' \
+    < "$lexicon"  >$tmpdir/g/select_empty.fst.txt
+  fstcompile --isymbols=$outdir/words.txt --osymbols=$outdir/words.txt $tmpdir/g/select_empty.fst.txt | \
+    fstarcsort --sort_type=olabel | fstcompose - $outdir/G.fst > $tmpdir/g/empty_words.fst
+  fstinfo $tmpdir/g/empty_words.fst | grep cyclic | grep -w 'y' && 
+  echo "Language model has cycles with empty words" && exit 1
+  rm -r $tmpdir/g
+fi
 echo "Succeeded in creating reversed language model."
 rm -r $tmpdir
